@@ -3,12 +3,12 @@ import 'package:pip_services3_commons/pip_services3_commons.dart';
 class Dummy implements IStringIdentifiable, ICloneable {
   @override
   String id;
-  String key;
-  String content;
+  String? key;
+  String? content;
 
-  Dummy();
+  Dummy() : id = IdGenerator.nextLong();
 
-  Dummy.from(this.id, this.key, this.content);
+  Dummy.from(String? id, this.key, this.content) : id = IdGenerator.nextLong();
 
   Map<String, dynamic> toJson() {
     return <String, dynamic>{'id': id, 'key': key, 'content': content};
